# contracts/openapi.yaml
openapi: 3.1.0
...
info:
  title: Accounting Platform API
  version: 1.0.0
  description: |
    Multi-tenant accounting platform. All business resources are tenant-scoped and most are
    legal-entity scoped; scope is taken from the authenticated session, never from a request
    parameter. Money is carried as decimal strings, never JSON numbers (ADR-0006).
  x-supersedes:
    - project plan/blueprints/api_route_catalog.yaml
servers:
  - url: https://api.example.com/v1
    description: Production
  - url: https://sandbox.api.example.com/v1
    description: Sandbox (doc 15 developer platform)
security:
  - sessionBearer: []
  - oauth2: []
...
paths:
...
  /reports/trial-balance:      {get: {operationId: getTrialBalance,       tags: [Reports], x-permission: report.view, x-idempotency: n-a, x-phase: 2, parameters: [{$ref: '#/components/parameters/BookId'}, {$ref: '#/components/parameters/PeriodFrom'}, {$ref: '#/components/parameters/PeriodTo'}], responses: {'200': {$ref: '#/components/responses/Ok'}, default: {$ref: '#/components/responses/Error'}}}}
...
  /reports/profit-loss:        {get: {operationId: getProfitLoss,         tags: [Reports], x-permission: report.view, x-idempotency: n-a, x-phase: 6, responses: {'200': {$ref: '#/components/responses/Ok'}, default: {$ref: '#/components/responses/Error'}}}}
  /reports/balance-sheet:      {get: {operationId: getBalanceSheet,       tags: [Reports], x-permission: report.view, x-idempotency: n-a, x-phase: 6, responses: {'200': {$ref: '#/components/responses/Ok'}, default: {$ref: '#/components/responses/Error'}}}}
  /reports/cash-flow:          {get: {operationId: getCashFlow,           tags: [Reports], x-permission: report.view, x-idempotency: n-a, x-phase: 6, responses: {'200': {$ref: '#/components/responses/Ok'}, default: {$ref: '#/components/responses/Error'}}}}
...
  /reports/drill-down: {get: {operationId: drillDownToJournalLines, tags: [Reports], x-permission: report.drill_down, x-idempotency: n-a, x-phase: 6, x-added-by: F-1005, description: "Phase 6 exit criterion 2 is \"every financial report amount can drill to ledger lines\", and doc 12 spells the chain out: statement total -> report row -> account/dimension breakdown -> journal lines -> source document. The last hop had no route of its own, so callers reached it through /reports/general-ledger, which caps at 5000 rows, takes no cursor and returns no marker - a drill on a control account with more lines than that answered with a prefix of the evidence and nothing said so. A truncated explanation of a total is worse than a refused one, because it looks like an answer. This route is the drill itself: cursor-paged, and its response carries has_more so a partial explanation is distinguishable from a complete one. It takes the same parameters as the report cell it came from, plus the report definition row when the cell is a range or a formula rather than a single account. Its own permission because doc 12 requires the permission on the destination record to be rechecked - \"a user may see an aggregate but not restricted source detail if policy requires\" - which is not expressible while the aggregate and the lines behind it share report.view (F-921).", parameters: [{$ref: '#/components/parameters/BookId'}, {$ref: '#/components/parameters/PeriodFrom'}, {$ref: '#/components/parameters/PeriodTo'}, {name: account_id, in: query, required: false, schema: {type: string, format: uuid}}, {name: report_definition_id, in: query, required: false, schema: {type: string, format: uuid}, description: "With row_no, drills a report row whose amount spans an account range or a formula."}, {name: row_no, in: query, required: false, schema: {type: integer}}, {name: dimension_value_id, in: query, required: false, schema: {type: string, format: uuid}}, {name: financial_snapshot_id, in: query, required: false, schema: {type: string, format: uuid}, description: "Drills an issued statement at the ledger cutoff it was taken at, rather than at now."}, {$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/Limit'}], responses: {'200': {$ref: '#/components/responses/JournalLinePage'}, default: {$ref: '#/components/responses/Error'}}}}
  /report-definitions:
    get:  {operationId: listReportDefinitions,  tags: [Reports], x-permission: report.view,   x-idempotency: n-a,         x-phase: 6, x-added-by: F-106, responses: {'200': {$ref: '#/components/responses/Ok'},      default: {$ref: '#/components/responses/Error'}}}
    post: {operationId: createReportDefinition, tags: [Reports], x-permission: report.manage, x-idempotency: recommended, x-phase: 6, x-added-by: F-106, description: "A change creates a new version; snapshots record the version used (doc 12).", responses: {'201': {$ref: '#/components/responses/Created'}, default: {$ref: '#/components/responses/Error'}}}
  /report-runs:
    get:  {operationId: listReportRuns,  tags: [Reports], x-permission: report.view, x-idempotency: n-a,         x-phase: 6, x-added-by: F-106, responses: {'200': {$ref: '#/components/responses/Ok'},      default: {$ref: '#/components/responses/Error'}}}
    post: {operationId: createReportRun, tags: [Reports], x-permission: report.export, x-idempotency: recommended, x-phase: 6, x-added-by: F-106, x-high-risk: true, description: "Asynchronous export job. doc 02 names \"export all financial data\" among the seven high-risk permissions, so report.export is flagged high-risk in the registry.", responses: {'202': {$ref: '#/components/responses/Accepted'}, default: {$ref: '#/components/responses/Error'}}}

  # ---------- Workflow (phase 6) ----------
...
  /audit-requests:
    get:  {operationId: listAuditRequests,  tags: [Audit], x-permission: audit.view,   x-idempotency: n-a,         x-phase: 6, responses: {'200': {$ref: '#/components/responses/Ok'},      default: {$ref: '#/components/responses/Error'}}}
    post: {operationId: createAuditRequest, tags: [Audit], x-permission: audit.manage, x-idempotency: recommended, x-phase: 6, responses: {'201': {$ref: '#/components/responses/Created'}, default: {$ref: '#/components/responses/Error'}}}
...
components:
  securitySchemes:
    sessionBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        First-party session token. Carries a session ID checked against the `sessions` table on
        refresh, so revocation is immediate (ADR-0005). Tenant is bound to the session, not the
        credential.
    oauth2:
      type: oauth2
      description: Third-party applications (doc 15). Scopes map 1:1 to x-permission values.
      flows:
        authorizationCode:
          authorizationUrl: https://api.example.com/oauth/authorize
          tokenUrl: https://api.example.com/oauth/token
          scopes: {}
  parameters:
...
    Cursor:     {name: cursor, in: query, required: false, schema: {type: string}, description: Opaque cursor; cursor pagination for all large collections.}
    Limit:      {name: limit,  in: query, required: false, schema: {type: integer, minimum: 1, maximum: 200, default: 50}}
    BookId:     {name: accounting_book_id, in: query, required: false, schema: {type: string, format: uuid}}
    PeriodFrom: {name: period_from, in: query, required: false, schema: {type: string, format: date}}
    PeriodTo:   {name: period_to,   in: query, required: false, schema: {type: string, format: date}}
...
  schemas:
    Money:
      type: string
      pattern: '^-?\d+(\.\d+)?$'
      description: Decimal string, never a JSON number (ADR-0006). Scale follows the currency's minor unit.
    Currency: {type: string, minLength: 3, maxLength: 3, description: ISO 4217}
    Uuid:     {type: string, format: uuid}
    Error:
      type: object
      required: [code, message, correlation_id]
      properties:
        code:           {type: string, description: Stable machine-readable code. Never a stack trace or SQL (doc 15).}
        message:        {type: string}
        field_errors:   {type: array, items: {type: object, properties: {field: {type: string}, code: {type: string}, message: {type: string}}}}
        correlation_id: {type: string}
    PageMeta:
      type: object
      properties:
        next_cursor: {type: string, nullable: true}
        has_more:    {type: boolean}
...
    JournalLine:
      type: object
      description: |
        One posted ledger line, as a drill returns it (F-1005). Amounts are decimal
        strings like every other monetary value (ADR-0006). `journal_entry_id`,
        `source_type` and `source_id` are the next two hops of doc 12's chain: a line
        must be followable to its entry and to the document that produced it, or the
        drill stops one step short of the evidence an auditor actually asks for.
      required: [id, journal_entry_id, posting_date, account_id, debit, credit]
      properties:
        id:                   {$ref: '#/components/schemas/Uuid'}
        journal_entry_id:     {$ref: '#/components/schemas/Uuid'}
        entry_number:         {type: string, description: "Allocated at posting and never reused (ADR-0004). A drill returns only POSTED and REVERSED lines, so it is always present; the id is what a link uses."}
        posting_date:         {type: string, format: date}
        account_id:           {$ref: '#/components/schemas/Uuid'}
        account_code:         {type: string}
        account_name:         {type: string}
        line_description:     {type: string, nullable: true}
        debit:                {$ref: '#/components/schemas/Money'}
        credit:               {$ref: '#/components/schemas/Money'}
        transaction_currency: {$ref: '#/components/schemas/Currency'}
        transaction_debit:    {$ref: '#/components/schemas/Money'}
        transaction_credit:   {$ref: '#/components/schemas/Money'}
        source_type:          {type: string, nullable: true, description: "The document type that produced the line - INVOICE, VENDOR_BILL, DEPRECIATION_RUN and so on. Null for a manual journal, which is its own source."}
        source_id:            {$ref: '#/components/schemas/Uuid'}
        dimensions:           {type: array, items: {type: object, properties: {dimension_id: {$ref: '#/components/schemas/Uuid'}, dimension_value_id: {$ref: '#/components/schemas/Uuid'}}}}
...
  responses:
    Ok:        {description: Success,  content: {application/json: {schema: {type: object}}}}
    Created:   {description: Created,  content: {application/json: {schema: {type: object}}}}
    Accepted:  {description: Accepted for asynchronous processing, content: {application/json: {schema: {type: object}}}}
...
    JournalLinePage:
      description: |
        Cursor page of posted journal lines behind one report amount.

        `meta.has_more` is REQUIRED on this response, not optional as it is on PageMeta
        (F-1005). A drill that returns fewer lines than exist must say so: an
        explanation of a total that has been silently trimmed reads exactly like a
        complete one, and the total it is explaining still balances either way. When
        `has_more` is true, `next_cursor` is non-null and the caller can obtain the
        rest; a page is never trimmed to fit a cap with neither marker set.
      content:
        application/json:
          schema:
            type: object
            required: [data, meta]
            properties:
              data: {type: array, items: {$ref: '#/components/schemas/JournalLine'}}
              meta:
                allOf: [{$ref: '#/components/schemas/PageMeta'}]
                required: [has_more, next_cursor]
    Error:
      description: |
        Structured error with a stable code and correlation ID. Notable codes:
        DUPLICATE_SUSPECTED (409, vendor bill override), IN_PROGRESS (409, idempotency),
        PERIOD_CLOSED (422), APPROVAL_REQUIRED (403), UNBALANCED_JOURNAL (422),
        OVER_ALLOCATION (422), INSUFFICIENT_STOCK (422).
      content: {application/json: {schema: {$ref: '#/components/schemas/Error'}}}